+++ app/admin.py
@@ -1,3 +1,6 @@
+from app.models import *
 from django.contrib import admin
 
 # Register your models here.
+admin.site.register(Pregunta)
+admin.site.register(Respuesta)
