+++ datos/admin.py
@@ -2,6 +2,4 @@ from datos.models import *
 from django.contrib import admin
 
 # Register your models here.
-admin.site.register(Pregunta)
-admin.site.register(Respuesta)
 admin.site.register(Usuario)
