+++ datos/forms.py
@@ -2,6 +2,6 @@ from .models import Usuario
 from django.contrib.auth.forms import UserCreationForm
 
 class SignUpForm(UserCreationForm):
-   class Meta:
-      model = Usuario
-      fields = ('email', 'username', 'password1', 'password2')
+  class Meta:
+    model = Usuario
+    fields = ('email', 'username', 'password1', 'password2')
